allow changing of dtypes to different bit width, separate compile-time and runtime data types

+++ skriptnd-pyproject/skriptnd/cpp/include/core/types.h
@@ -27,7 +27,7 @@ namespace sknd
 {
     
     typedef float real_t;
-    typedef int int_t;
+    typedef long int_t;
     typedef bool bool_t;
     typedef std::string str_t;
